Much more linux progress (won't compile yet)

+++ Pongbot/lib/hlsdk/public/tier0/memalloc.h
@@ -178,7 +178,8 @@ inline void *MemAlloc_AllocAligned( size_t size, size_t align )
 {
 	unsigned char *pAlloc, *pResult;
 
-	if (!IsPowerOfTwo(align))
+	// Add cast - pongo1231
+	if (!IsPowerOfTwo((int) align))
 		return NULL;
 
 	align = (align > sizeof(void *) ? align : sizeof(void *)) - 1;
@@ -196,7 +197,8 @@ inline void *MemAlloc_AllocAligned( size_t size, size_t align, const char *pszFi
 {
 	unsigned char *pAlloc, *pResult;
 
-	if (!IsPowerOfTwo(align))
+	// Add cast - pongo1231
+	if (!IsPowerOfTwo((int) align))
 		return NULL;
 
 	align = (align > sizeof(void *) ? align : sizeof(void *)) - 1;
@@ -248,7 +250,8 @@ inline void *MemAlloc_AllocAlignedFileLine( size_t size, size_t align, const cha
 
 inline void *MemAlloc_ReallocAligned( void *ptr, size_t size, size_t align )
 {
-	if ( !IsPowerOfTwo( align ) )
+	// Add cast - pongo1231
+	if (!IsPowerOfTwo((int) align))
 		return NULL;
 
 	// Don't change alignment between allocation + reallocation.
